BUG: cast to a good print type.
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=4551
git-svn-id: http://svn.slicer.org/Slicer4/trunk@4551 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/ModuleDescriptionParser/ModuleProcessInformation.cxx
@@ -16,7 +16,7 @@
 
 std::ostream& operator<<(std::ostream &os, const ModuleProcessInformation &p)
 {
-  os << "  Abort: " << p.Abort << std::endl;
+  os << "  Abort: " << (short) p.Abort << std::endl;
   os << "  Progress: " << p.Progress << std::endl;
   os << "  Stage Progress: " << p.StageProgress << std::endl;
   os << "  ProgressMessage: " << p.ProgressMessage << std::endl;
